Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/security.drawImage.image.cross-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
security.drawImage.image.cross
drawImage of different-origin image makes the canvas origin-unclean
Actual output:

PASS drawImage of different-origin image makes the canvas origin-unclean
